footer-component

+++ src/app/cart-home/cart-home.component.ts
@@ -4,8 +4,9 @@ import { Product } from '../models/product';
 import { CatalogComponent } from '../catalog/catalog.component';
 import { ItemCart } from '../models/itemCart';
 import { HeaderComponent } from '../header/header.component';
-import { RouterOutlet } from '@angular/router';
+import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../services/sharing-data.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cart-home',
@@ -22,8 +23,11 @@ export class CartHomeComponent implements OnInit {
   //price total
   total: number = 0;
 
-
-  constructor(private sharingDataService: SharingDataService ,private service: ProductService) {}
+  constructor(
+    private router: Router,
+    private sharingDataService: SharingDataService,
+    private service: ProductService
+  ) {}
   ngOnInit(): void {
     this.product = this.service.findAll();
     /*usamos el operador ( ! ) para determinar que no sea vacio*/
@@ -31,12 +35,10 @@ export class CartHomeComponent implements OnInit {
     this.calculateTotal();
     this.deleteCart();
     this.onAddCart();
-
   }
 
   onAddCart() {
-    this.sharingDataService.productEventEmitter.subscribe( product =>{
-
+    this.sharingDataService.productEventEmitter.subscribe((product) => {
       const hasItem = this.items.find((item) => {
         return item.product.id === product.id;
       });
@@ -56,26 +58,54 @@ export class CartHomeComponent implements OnInit {
       }
       this.calculateTotal();
       this.saveSession();
+      this.router.navigate(['/cart'], {
+        state: { items: this.items, total: this.total },
+      });
 
-
-    })
-
+      Swal.fire({
+        title: 'Agregado a tu Carrito',
+        text: 'Ya puedes revisar tus productos',
+        icon: 'success',
+      });
+    });
   }
 
   deleteCart(): void {
-    this.sharingDataService.idProductEventEmitter.subscribe(id =>{
-      this.items = this.items.filter((item) => item.product.id !== id);
-      if(this.items.length == 0){
-        sessionStorage.removeItem('cart')
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
-    })
+    this.sharingDataService.idProductEventEmitter.subscribe((id) => {
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.items = this.items.filter((item) => item.product.id !== id);
+          if (this.items.length == 0) {
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+          this.calculateTotal();
+          this.saveSession();
 
+          this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
+            this.router.navigate(['/cart'], {
+              state: { items: this.items, total: this.total },
+            });
+          });
+          Swal.fire({
+            title: 'Deleted!',
+            text: 'Your file has been deleted.',
+            icon: 'success',
+          });
+        }
+      });
+
+    });
   }
   //reducir el flujo a un solo valor
-
 
   calculateTotal(): void {
     this.total = this.items.reduce(
@@ -87,5 +117,4 @@ export class CartHomeComponent implements OnInit {
   saveSession(): void {
     sessionStorage.setItem('cart', JSON.stringify(this.items));
   }
-
 }
